requirements.txt
# FastAPI 核心依赖
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3,<3.0.0

# Claude API
anthropic>=0.40.0  # 更新最低版本要求
mcp>=1.0.0       # MCP 1.x 版本

# OpenAI SDK（用于千问等兼容接口）
openai>=1.0.0      # 🆕 千问使用 OpenAI 兼容接口

# Qwen DashScope SDK
dashscope>=1.25.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 其他工具
python-multipart>=0.0.6
aiofiles>=24.1.0          # 异步文件操作
aiohttp>=3.9.0            # 异步 HTTP 客户端
httpx>=0.27.0     # HTTP 客户端（用于 MCP、测试）
filetype>=1.2.0           # 文件类型检测（纯 Python，无需系统依赖）

# 数据库（PostgreSQL）
asyncpg>=0.29.0           # PostgreSQL 异步驱动
greenlet>=3.0.0           # SQLAlchemy 异步支持
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0           # 数据库迁移

# Redis（Session 管理和 SSE 重连）
redis>=5.0.0,<6.0.0

# Token 计数
tiktoken>=0.5.0

# JSON 解析（宽松模式）
json5>=0.9.0

# E2B Sandbox
# 注意：需要 v2.x 才能使用 beta_create 和 auto_pause 功能
e2b>=2.0.0
e2b-code-interpreter>=2.0.0

# AWS S3 存储（异步）
aioboto3>=13.0.0
boto3>=1.34.0  # aioboto3 依赖 boto3

# gRPC（内部微服务通信）- 版本需要匹配
grpcio>=1.66.0
grpcio-tools>=1.66.0
grpcio-reflection>=1.66.0
protobuf>=5.26.0

# Mem0 记忆系统
mem0ai>=0.1.0
tcvectordb>=1.0.0         # 腾讯向量数据库

# 文件处理
Pillow>=10.0.0            # 图片压缩
PyPDF2>=3.0.0             # PDF 文本提取
python-docx>=0.8.11       # Word 文档解析（页数估算）
python-pptx>=0.6.21       # PPT 生成和解析
openpyxl>=3.0.0           # Excel 表格解析（工作表数量）
odfpy>=1.4.1              # ODT 文档解析
unstructured-client>=0.30.6  # 非结构化数据处理客户端

# 定时任务
apscheduler>=3.10.0

# Google Gemini API（图像生成）
google-genai>=1.0.0       # Gemini 3 Pro Image API

# 测试工具
pytest>=8.0.0             # 测试框架
pytest-asyncio>=0.23.0    # 异步测试支持
pytest-dotenv>=0.5.2      # 自动加载 .env 文件
# 系统监控
psutil>=5.9.0

# WebSocket（实时通信）
websockets>=12.0
